Update Developer-Guide_Using-Vagrant.md
Igor,  this is my stab at the documentation for building with Vagrant as the system is now,  without using the llnux specific features (i.e. installing vagrant and virtualbox from compile.sh)   Please feel free to edit as you see fit.  I did just run through this from doing a clean install of vagrant on my machine. 

 There does seem to be a problem with the dialog package not only not being installed in Ubuntu,  but also not being installable.  I downloaded the package from one of the repositories  and installed it myself in the vagrant machine.  A few days ago,  it wasn't available for ./compile.sh but I was able to use apt to install it in the Vagrant machine.   I'm assuming this is some kind of virtual box cloud version glitch.  I know that the package is supposed to be available becauase I am able to install it on my Ubuntu focal .

I don't have a mac on which to test.